efficientdet-*-tf: make conversion work when Keras is installed
The models' sources contain their own `keras` package, and without the
`__init__.py` file it's not recognized as a package, with the system `keras`
package getting imported instead. (It works when there's no system `keras`
package, because then it's recognized as a namespace package).

Download `keras/__init__.py` to make it always work.