demos: wait for every callback
inference_completion_callback() could be executed when main() is already
considered finished. That resulted in segfault. Make sure every callback
is processed. The same problem was for C++. The hypothesis is C++ is
fast enough to finish all callbacks while main() finalizes and prints
statistics, so the problem didn't appear.
Another possible solution is to override callbacks with empty ones on
exit. But since demos follow the approach of processing leftovers, every
callback waiting approach is taken.
It makes sense to implement AsyncPipeline as a context manager anyway
and stop all callbacks on __exit__() to handle exceptions risen from
main() properly. I don't do it here. The aim for now is to enable
CPU_THROUGHPUT_AUTO having precommits passed, plus in the very end I
would prefer wait() based approach, not callback based one, handling
work with callbacks in a separate entity like how it is done for Python
in IE.