tools/downloader: put the model tools into a Python package
A package is needed so that we can make them installable via pip later.
I don't think you can install individual modules (and even if you could,
the names `downloader`, etc., are way too generic and would pollute the global
namespace).

I also used this refactoring to rename `common` to `_common`, to signify that
it's an internal module.

`open_model_zoo` is a namespace package, so that if we end up creating other
installable OMZ functionality (like a model API or something), we can put it
in a different subpackage of `open_model_zoo`. `open_model_zoo.model_tools`,
however, is not a namespace package, since there should be no reason to extend
it.