tools/downloader: dump the information about which models are quantizable
This has been long overdue - now the users of the info dumper can determine
programmatically which models can be quantized.

Instead of a boolean, we report the quantizer output precisions. This way,
we can have different precisions per model in the future. Similarly for
the added `if` in the quantizer - it's currently useless, as there's no way
to skip every precision for a model, but if some model has an unusual set
of output precisions in the future, that will become a possibility.

The name "quantization_output_precisions" is a bit bulky, but I thought that
if it was called just "quantization_precisions", it would be unclear whether
that was referring to input or output precisions.