Model Converter: try to run Model Optimizer as a package
If our tools are installed as part of the `openvino-dev` distribution, then
it doesn't make sense to ask the user to run `setupvars`, seeing as:

1. It doesn't exist.
2. We know that MO is installed as part of the same distribution.

Therefore, before trying to find MO the old way, try to locate its package
and if that succeeds, run it as a package.

I moved the documentation for the `-p` option up, because I'm
referencing its value in the documentation for the `--mo` option.

`get_package_path` is in `_common`, because we're likely going to implement
the same thing in the quantizer (for POT).