Model Converter: move the convert function to the module level
`convert` initially started small, so it made sense to make it a local
function within `main`. But then it grew and grew, to the point where it's now
the same length as the rest of `main` (not counting the argument parsing).
It's clearly no longer readable to keep it as a local function, so move it
to the module level instead.