Work around Layout == str issue (#3358)
The Layout utility datatype was recently introduced. However, it seems
to be polymorphic depending on the way it was constructed: sometimes
it's just a string, sometimes it has a data member called "layout".

This is just a fix to make OVMSAdapter work with the most common use
case. There is a bigger fix needed to patch all code paths in
OVMSAdapter and OpenvinoAdapter.