Fix ModelAPI setup.py
I have checked what it takes to create a PyPi package. We already have
setup.py which can generate needed artifacts. We only need to rearrange
the folders structure to remove the openvino folder. So if OTX changes
their mind and refuses to use OMZ as a submodule, we should be able to
publish ModelAPI to PyPi. But I guess we will need to use my personal
account instead of Intel’s one for publishing since we won’t follow
common processes.