Beautify workflows
I noticed, that after a merge, GitHub Actions builds are named just “build”. It turns out this is because we had keys in `jobs` map named that way. Giving each job a meaningful name removes a need to have multiple .yml files with their names.

Add strategy matrix for accuracy_checker simply because it’s easy.